(Started to) add some instances to the BliSet typeclass.
Note that I think our current interface is too general, and needs to be revised. For instance,
our collection for the schema should not just implement efficent lookups of schema entries, but
also permit efficently telling if a value with a certain arity (or, with a predicate matching
a certain arity) exists in our schema. And in that case we need a more general "lookup" as part of
the interface.

Also, the datatype that is the most efficent for our collection of schema entries, and the datatype
that is the most efficent for our collection of facts may not be the same, so our "generic" monads
should actually be generic over both of these container types.

Moreover, the interface for these two containers may in fact be different.

Thus, before I try to move the main code base into this generic interface, I should
think more about what the interface for these two container types should be.